multi file changes 2
-working copy of multifile done for Basic Texture Slicing Effect.
-updated common.h and common.cpp
-updated code into new Header and cpp files inside \Basic_TextureSlicing\ and \utils\ and \OpenGL\
-updated build.bat file  accordingly

TODO: code cleanup for removing commented code is needed
