+++ build.bat
@@ -1,14 +1,31 @@
 
+cls
+
 del VolumeApp.exe
 del VolumeApp.obj
 del OGL.res
+del ".\\bin\\*.obj"
 
 
-cl.exe /c /EHsc /I ".\\dependencies\\glew\\include\\GL" VolumeApp.cpp
+@REM cl.exe /c /EHsc /I ".\\dependencies\\glew\\include\\GL"  VolumeApp.cpp
 
-rc.exe OGL.rc
+@REM Compiles all the available cpp files in the subdirectories
+for /r %%i in (*.cpp) do (
+	echo "%%i"
+    cl.exe /Fo".\\bin\\" /c /EHsc  /I ".\\dependencies\\glew\\include\\GL" "%%i"
+)
 
-link.exe VolumeApp.obj OGL.res user32.lib gdi32.lib /LIBPATH:".\\dependencies\\glew\\lib\\Release\\x64" /SUBSYSTEM:WINDOWS
+echo ====================== COMPILATION DONE ======================
+
+@REM rc.exe OGL.rc
+rc.exe /Fo".\\bin\\OGL.res" OGL.rc
+
+@REM link.exe VolumeApp.obj OGL.res user32.lib gdi32.lib /LIBPATH:".\\dependencies\\glew\\lib\\Release\\x64" /SUBSYSTEM:WINDOWS
+
+
+link.exe /out:VolumeApp.exe ./bin/*.obj ./bin/OGL.res user32.lib gdi32.lib /LIBPATH:".\\dependencies\\glew\\lib\\Release\\x64" /SUBSYSTEM:WINDOWS
+
+echo ====================== LINKING DONE ======================
 
 VolumeApp.exe
