+++ src/utils/OpenGL/GLLog.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <windows.h> // For BOOL, TEXT(), MessageBox(), etc.
+#include <stdio.h> // For fopen_s(), fprintf(), fclose()
+
+#include "GLHeadersAndMacros.h"
+
+
+BOOL CreateLogFile(void);
+void PrintGLInfo(void);
+void PrintLog(const char *, ...);
+void CloseLogFile(void);
+
